Add files via upload
A stand alone version of tetris that works off a single micro:bit.  It is the server code but stripped of all the matrix functionality.  It prints each frame to REPL... falling blocks are shown with '7', static blocks (that have already reached their place of rest) are shown as '1' and blanks are '0'.
Use A to move shape left, B to move it right
Tilt left and tilt right to rotate as you'd expect.